--- a/.taprc
+++ b/.taprc
@@ -1,1 +1,1 @@
-coverage: false
+coverage: true